refactor : ResultView와 LottoApplication(Controller) 분리
